Final stretch, game flow, tweaks/polish, balancing

+++ Assets/KJam/Game/Scripts/SaveInfo.cs
@@ -9,7 +9,7 @@ public struct SaveInfo
 {
 	public string BuildVersion; // For jam updates
 	public int LevelsPlayed; // For level progression stuff.. just jam things
-	public float Gold;
+	public int Gold;
 	public string[] Items;
 	public string[] EquippedItemsKey;
 	public int[] EquippedItemsValue;
